+++ assets/js/hooks/Mapper/components/ui-kit/index.ts
@@ -23,3 +23,4 @@ export * from './MenuItemWithInfo';
 export * from './MarkdownTextViewer.tsx';
 export * from './WdButton.tsx';
 export * from './constants.ts';
+export * from './RespawnTag';